feat: introduce unit tests
